requirements.txt
openai
flask
python-dotenv
xmltodict
requests>=2.33.0
prometheus-client
python-json-logger
defusedxml
pytest
maxminddb
python-nmap
atheris
<<<<<<< HEAD

# Hybrid RAG dependencies (v0.4.3)
langchain>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.0

# REST API dependencies (v0.8.0)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic>=2.0.0
=======
azure-ai-inference==1.0.0b9
numpy>=1.24.0
scikit-learn>=1.3.0
>>>>>>> origin/main
